EHP Solutions/Kafka Healthtrixss Architecture/server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
kafka-python==2.0.2
redis==4.5.4
pandas==1.5.3
pydantic==1.10.7
pytest==7.3.1
requests==2.31.0
python-dotenv==1.0.0